+++ Vaango/src/CCA/Components/MPM/ParticleCreator/ShellParticleCreator.h
@@ -80,7 +80,7 @@ WARNING
     // Actually create particles using geometry
     //
     virtual particleIndex createParticles(MPMMaterial* matl, 
-                                            CCVariable<short int>& cellNAPID,
+                                            CCVariable<int>& cellNAPID,
                                             const Patch*, 
                                             DataWarehouse* new_dw,
                                             std::vector<GeometryObject*>&);
